address review: extract shared chat-completions stub harness for retry suites
